# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoroTrail API",
    "description": "A contract event indexer for the Stellar/Soroban network. SoroTrail polls\na Stellar RPC endpoint, stores contract events durably in Postgres, and\nserves them back through a queryable HTTP API long after the RPC has\nforgotten them.\n",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "TenantAuth": []
    },
    {
      "TenantAuthHeader": []
    },
    {}
  ],
  "paths": {
...
    "/events": {
      "get": {
        "operationId": "listEvents",
        "summary": "List stored events",
        "description": "Paginated list of stored events, filterable by contract, type, topic,\nledger range, time range, and sort order. Supports cursor-based\npagination and optional spec-driven enrichment.\n",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractID"
          },
          {
            "$ref": "#/components/parameters/EventType"
          },
          {
            "$ref": "#/components/parameters/Topic"
          },
          {
            "$ref": "#/components/parameters/Topic0"
          },
          {
            "$ref": "#/components/parameters/Topic1"
          },
          {
            "$ref": "#/components/parameters/Topic2"
          },
          {
            "$ref": "#/components/parameters/Topic3"
          },
          {
            "$ref": "#/components/parameters/FromLedger"
          },
          {
            "$ref": "#/components/parameters/ToLedger"
          },
          {
            "$ref": "#/components/parameters/FromTime"
          },
          {
            "$ref": "#/components/parameters/ToTime"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Order"
          },
          {
            "$ref": "#/components/parameters/Decoded"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            }
          },
          "304": {
            "$ref": "#/components/responses/NotModified"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "operationId": "deleteEvents",
        "summary": "Prune stored events",
        "description": "Deletes stored events matching the given bounds. Intended for retention maintenance; gated behind the management credential.",
        "tags": [
          "Operations"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ToLedger"
          }
        ],
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Number of events deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "operationId": "getEvent",
        "summary": "Get a single event by ID",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "TOID-based event identifier",
            "schema": {
              "type": "string",
              "pattern": "^[0-9]{16}-[0-9]{10}$"
            },
            "example": "0001099511627776-0000000001"
          },
          {
            "$ref": "#/components/parameters/Decoded"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested event",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Event"
                    },
                    {
                      "$ref": "#/components/schemas/EnrichedEvent"
                    }
                  ]
                }
              }
            }
          },
          "304": {
            "$ref": "#/components/responses/NotModified"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/contracts/{id}/events": {
      "get": {
        "operationId": "contractEvents",
        "summary": "List events for a contract",
        "description": "Convenience wrapper around GET /events with contract_id={id}",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Stellar contract ID (C-prefixed)",
            "schema": {
              "type": "string",
              "pattern": "^C[A-Z0-9]{55}$"
            },
            "example": "CDLZFC3SYJYDZT7K67VZ75HPJVIEUVNIXF47ZG2FB2RMQQVU2HHGCYSC"
          },
          {
            "$ref": "#/components/parameters/EventType"
          },
          {
            "$ref": "#/components/parameters/Topic"
          },
          {
            "$ref": "#/components/parameters/Topic0"
          },
          {
            "$ref": "#/components/parameters/Topic1"
          },
          {
            "$ref": "#/components/parameters/Topic2"
          },
          {
            "$ref": "#/components/parameters/Topic3"
          },
          {
            "$ref": "#/components/parameters/FromLedger"
          },
          {
            "$ref": "#/components/parameters/ToLedger"
          },
          {
            "$ref": "#/components/parameters/FromTime"
          },
          {
            "$ref": "#/components/parameters/ToTime"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Order"
          },
          {
            "$ref": "#/components/parameters/Decoded"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of events for the contract",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            }
          },
          "304": {
            "$ref": "#/components/responses/NotModified"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/addresses/{address}/events": {
      "get": {
        "operationId": "addressEvents",
        "summary": "List events involving an address",
        "description": "Events where the given Stellar address appears in the event topics, as either sender or recipient. Backed by the address activity index.",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "description": "Stellar account (G-prefixed) or contract (C-prefixed) address",
            "schema": {
              "type": "string"
            },
            "example": "GDW6AUTBXTOC7FIKUO5BOO3OGLK4SF7ZPOBLMQHMZDI45J2Z6VXRB5NR"
          },
          {
            "$ref": "#/components/parameters/FromLedger"
          },
          {
            "$ref": "#/components/parameters/ToLedger"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Order"
          }
        ],
        "responses": {
          "200": {
            "description": "A page of events involving the address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/addresses/{address}/summary": {
      "get": {
        "operationId": "addressSummary",
        "summary": "Summarise an address's event history",
        "description": "Aggregate view of an address's activity: first and last ledger it was seen in, total event count, and the contracts it interacted with.",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "description": "Stellar account (G-prefixed) or contract (C-prefixed) address",
            "schema": {
              "type": "string"
            },
            "example": "GDW6AUTBXTOC7FIKUO5BOO3OGLK4SF7ZPOBLMQHMZDI45J2Z6VXRB5NR"
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregate activity for the address",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/admin/tenants": {
      "get": {
        "operationId": "listTenants",
        "summary": "List tenants",
        "description": "Administrative endpoint, available only when MULTI_TENANT is enabled and gated behind the admin credential.",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "operationId": "createTenant",
        "summary": "Create a tenant",
        "description": "Administrative endpoint, available only when MULTI_TENANT is enabled and gated behind the admin credential.",
        "tags": [
          "Admin"
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "Topic0": {
        "name": "topic0",
        "in": "query",
        "description": "Exact JSON match against topic position 0",
        "schema": {
          "type": "string"
        },
        "example": "{\"symbol\":\"transfer\"}"
      },
...
    },
...
    "securitySchemes": {
      "TenantAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The tenant API key issued by POST /admin/tenants/{id}/keys, presented\nas `Authorization: Bearer st_<prefix>_<secret>`. The same key is also\naccepted in an `X-API-Key` header (see TenantAuthHeader); Bearer wins\nwhen both are present.\n\nOnly enforced when the instance runs with MULTI_TENANT=true. On a\nsingle-tenant instance every request is served as an implicit\nwildcard principal and no credential is required, which is why most\noperations list this scheme as optional rather than required.\n"
      },
      "TenantAuthHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "The same tenant API key as TenantAuth, presented in a header instead of a Bearer token."
      },
      "ManagementKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "The operator's instance-wide management key, configured through the\nAPI_KEY environment variable. It gates the endpoints that mutate\ningestion or read the dead-letter queue, and it is always required on\nthose endpoints — independently of MULTI_TENANT, and independently of\nthe tenant credential above.\n\nWhen API_KEY is unset the server fails closed: those endpoints answer\n503 rather than running unauthenticated.\n"
      }
    },
...
  }
}